Refactor node/msv computation
